validate/requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0
matplotlib>=3.7.0
Pillow>=10.0.0
pandas>=2.0.0
PyYAML>=6.0
tqdm>=4.65.0
hypothesis>=6.82.0
requests>=2.31.0
wget>=3.2
imageio>=2.31.0
